*: use uaccess instead of plugdev
The plugdev group is something Ubuntu/Debian use to grant access to
devices. The uaccess mechanism ships with systemd and uses ACLs to
control access instead. This mechanism is advantageous because it only
grants access to sessions that are seated, as opposed to any session at
all (e.g. SSH).

This also adds the udev rule needed for flashing the Moonlander.