deepwater: switch back to gnome
That was a neat experiment. This doesn't ultimately seem to be any
faster. I think the rapid transitions certainly make it feel that way
(and the animations _are_ actually slower than instant), but overall,
this doesn't seem to be tremendously better than gnome. I never bothered
to set up a lockscreen either, and I'm lazy enough that I'm probably not
going to figure it out. I like the extensibility of the system - the
concept of panels is pretty elegant - but I find the plugins to be ugly
and the battery meter doesn't stand out enough when the battery is low.