tacoma: enable dmz bridge
This will allow multiple VMs to use the same VLAN device. Without the
bridge, only a single macvtap could be created by libvirt, allowing only
a single machine to use the interface. Creating a bridge and connecting
the VMs to that interface instead allows for many VMs to share the
underlying VLAN device.