matrix: enable mautrix-signal
This will not automatically work on a fresh install. Mautrix-signal
will need to be registered before, the permissions on
`/var/lib/signald/attachments` modified to allow the signald group to
write (so that mautrix-signal may delete attachments), and potentially
other steps that I'm forgetting - this took a lot of trial and error.