tacoma: automatically start plex reliably
For longer than I care to admit, Plex has been failing to start after
boot with the following error:

```
tacoma systemd[1]: plex.service: Job plex.service/start failed with
                   result 'dependency'.
tacoma systemd[1]: Dependency failed for Plex Media Server.
```

This is due to a long standing systemd bug [1] which prevents services
from being started if any of their dependency fail to start - even if
they have a restart policy.

A workaround I found recently suggested that using the Upholds
directive in a mount unit would allow a service to be automatically
started once that mount was available. Custom mount directives cannot
be defined via fstab entries however, so I tried redefining the mount
using a systemd mount directly. This ended up not working though
because of the use of the credentials mount option, which isn't
respected by the kernel (probably so that it doesn't have to read
anything from the userspace filesystem). As it turns out, the
`mount.cifs` utility reads the credentials file and inserts the
username and password into its mount syscall, but this utility isn't
present for `mount -t cifs` to discover. To get this to work, I just
duplicated the mount between both the fstab entry and systemd unit.

[1]: https://github.com/systemd/systemd/issues/1312