matrix: ensure mautrix-signal is started after signald
After having successfully used the signal bridge for a while, I
noticed I was suddenly unable to send attachments. An error like the
following was posted by the bridge bot after each attempt:

    Your message was not bridged: InternalError:
    java.io.FileNotFoundException: java.io.FileNotFoundException:
    /tmp/mautrix-signal-a5cc5ddf-23c8-454e-acb5-4431bb1484d0 (No such
    file or directory)

Poking around a bit revealed that mautrix-signal had started before
signald. In fact, the processes had sequentially adjacent PIDs!
Restarting mautrix-signal fixed the issue. Based on this, it would
seem that the `JoinsNamespaceOf` directive requires there to be an
ordering between the services, though it doesn't enforce it itself.