ddcci: add support for desktop backlight control
The ddcci module doesn't yet support monitor hot-plugging, so a udev
rule and systemd service are used to work around the issue. This was
adapted from a comment [1] in the bug tracker. The udev rule is
specific to the ATI card in Pepcon, so if this ends up being needed
elsewhere, it will need to be parameterized.

[1]: https://gitlab.com/ddcci-driver-linux/ddcci-driver-linux/-/issues/7#note_151296583