emacs: use home-manager and include init.el
This also no longer removes the Emacs (daemon) shortcut. I couldn't
quickly figure out how to keep that functionality, plus I think it was
causing builds to be done from source, so this probably works out. The
init.el is a bit of a mess and doesn't entirely work, but it's a start.